pyproject.toml
[tool.poetry]
name = "wlapp"
version = "0.1.0"
description = ""
authors = ["nagpalpranay <54006833+nagpalpranay@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
langchain = "^0.2.16"
langchain-openai = "^0.1.23"
langchain-qdrant = "^0.1.3"
langchain-community = "^0.2.16"
python-dotenv = "^1.0.1"
langgraph = "^0.2.19"
tavily-python = "^0.4.0"
pydantic = "^2.9.1"
langchain-groq = "^0.1.9"
langchain-experimental = "^0.0.65"
langchain-google-vertexai = "^1.0.10"
vertexai = "^1.67.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
